[T-4] Added posts and comments endpoints, permissions

+++ groups/views.py
@@ -1,11 +1,11 @@
 from django.utils.datastructures import MultiValueDictKeyError
 from rest_framework import viewsets
 from rest_framework.decorators import action
-from rest_framework.permissions import IsAuthenticated
+from rest_framework.pagination import PageNumberPagination
 
+from core.permissions import IsLecturerOrIsAdmin, IsOwnerOrIsModerator, IsOwner, IsMember, set_basic_permissions
 from core.responses import response200, response406, response404
 from groups.models import Group, PendingMember
-from groups.permissions import IsLecturerOrIsAdmin, IsOwnerOrModerator, IsOwner, IsMember
 from groups.serializers import GroupSerializer, PendingMembersSerializer
 from users.models import User
 
@@ -13,9 +13,21 @@ from users.models import User
 class GroupViewSet(viewsets.GenericViewSet):
     queryset = Group.objects.all()
 
+    @action(methods=['GET'], detail=False, url_name='search', url_path='search')
+    def search_for_groups(self, request):
+        try:
+            search_phrase = request.GET['phrase']
+        except MultiValueDictKeyError:
+            return response406({'message': 'Wrong search key'})
+        groups = Group.objects.filter(name__contains=search_phrase)
+        groups_data = GroupSerializer(groups, many=True).data
+        paginator = PageNumberPagination()
+        data = paginator.paginate_queryset(groups_data, request)
+        return paginator.get_paginated_response(data)
+
     @action(methods=['POST'], detail=False, url_name='create_group', url_path='create')
     def create_group(self, request):
-        serializer = GroupSerializer(data=request.data)
+        serializer = GroupSerializer(data=request.data, partial=True)
         if not serializer.is_valid():
             return response406({**serializer.errors, 'message': 'Invalid input data'})
         serializer.save(owner=request.user)
@@ -35,6 +47,7 @@ class GroupViewSet(viewsets.GenericViewSet):
             group = Group.objects.get(**kwargs)
         except Group.DoesNotExist:
             return response404('Group')
+        self.check_object_permissions(request=request, obj=group)
         serializer = GroupSerializer(group, data=request.data, partial=True)
         if not serializer.is_valid():
             return response406({**serializer.errors, 'message': 'Validation error'})
@@ -47,6 +60,7 @@ class GroupViewSet(viewsets.GenericViewSet):
             group = Group.objects.get(**kwargs)
         except Group.DoesNotExist:
             return response404('Group')
+        self.check_object_permissions(request=request, obj=group)
         group.delete()
         return response200({'message': 'Group has been successfully deleted'})
 
@@ -56,6 +70,7 @@ class GroupViewSet(viewsets.GenericViewSet):
             group = Group.objects.get(**kwargs)
         except Group.DoesNotExist:
             return response404('Group')
+        self.check_object_permissions(request=request, obj=group)
         group.members.remove(request.user)
         group.save()
         return response200({'message': 'You successfully left the group'})
@@ -69,6 +84,7 @@ class GroupViewSet(viewsets.GenericViewSet):
             return response404('Group')
         except MultiValueDictKeyError:
             return response406({'message': 'Wrong parameters'})
+        self.check_object_permissions(request=request, obj=group)
         group.members.remove(user)
         group.save()
         return response200({'message': 'Successfully dropped user from group'})
@@ -92,6 +108,7 @@ class GroupViewSet(viewsets.GenericViewSet):
             return response404('Group')
         except PendingMember.DoesNotExist:
             return response404('Pending member')
+        self.check_object_permissions(request=request, obj=group)
         group.members.add(pending.user)
         pending.delete()
         return response200({'message': 'Successfully accepted user'})
@@ -105,21 +122,28 @@ class GroupViewSet(viewsets.GenericViewSet):
             return response404('Group')
         except PendingMember.DoesNotExist:
             return response404('Pending member')
+        self.check_object_permissions(request=request, obj=group)
         pending.delete()
         return response200({'message': 'Successfully declined user'})
 
     def get_permissions(self):
-        lecturer_or_admin_actions = ['create_group']
-        owner_or_moderator_actions = ['update_group', 'drop_member', 'accept_pending_member', 'decline_pending_member']
-        group_owner_actions = ['delete_group']
-        group_member = ['leave_group']
-        self.permission_classes.append(IsAuthenticated)
-        if self.action in lecturer_or_admin_actions:
-            self.permission_classes = [*self.permission_classes, IsLecturerOrIsAdmin]
-        if self.action in owner_or_moderator_actions:
-            self.permission_classes = [*self.permission_classes, IsOwnerOrModerator]
-        if self.action in group_owner_actions:
-            self.permission_classes = [*self.permission_classes, IsOwner]
-        if self.action in group_member:
-            self.permission_classes = [*self.permission_classes, IsMember]
+        action_types = [
+            {
+                'class': IsOwnerOrIsModerator,
+                'values': ['update_group', 'drop_member', 'accept_pending_member', 'decline_pending_member']
+            },
+            {
+                'class': IsOwner,
+                'values': ['delete_group']
+            },
+            {
+                'class': IsLecturerOrIsAdmin,
+                'values': ['create_group']
+            },
+            {
+                'class': IsMember,
+                'values': ['leave_group']
+            }
+        ]
+        self.permission_classes = set_basic_permissions(self.action, action_types)
         return [permission() for permission in self.permission_classes]
